refactor paraspell files
