+++ app/src/hooks/useBalance.tsx
@@ -2,7 +2,7 @@ import { Chain } from '@/models/chain'
 import { Token } from '@/models/token'
 import { Erc20Balance } from '@/services/balance'
 import { Environment } from '@/store/environmentStore'
-import { getCurrencyId, getNativeToken, getParaSpellNode } from '@/utils/paraspell'
+import { getCurrencyId, getNativeToken, getParaSpellNode } from '@/utils/paraspellTransfer'
 import { toHuman } from '@/utils/transfer'
 import { getTransferableAmount, TNodeDotKsmWithRelayChains } from '@paraspell/sdk'
 import { captureException } from '@sentry/nextjs'
