+++ app/src/hooks/useFees.tsx
@@ -6,7 +6,7 @@ import { PolkadotTokens } from '@/registry/mainnet/tokens'
 import { getCachedTokenPrice } from '@/services/balance'
 import { Direction, resolveDirection } from '@/services/transfer'
 import { getPlaceholderAddress } from '@/utils/address'
-import { getCurrencyId, getNativeToken, getParaSpellNode } from '@/utils/paraspell'
+import { getCurrencyId, getNativeToken, getParaSpellNode } from '@/utils/paraspellTransfer'
 import { getRoute, isSameChain } from '@/utils/routes'
 import { getFeeEstimate } from '@/utils/snowbridge'
 import { toHuman } from '@/utils/transfer'
